% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjCombos.R
\name{adjCombos}
\alias{adjCombos}
\title{Create Adjacent Combinations, Varying Length}
\usage{
adjCombos(invec)
}
\arguments{
\item{invec}{The input vector}
}
\value{
A \code{list} of vectors
}
\description{
Create adjacent combinations of the elements of a vector, varying the length
with each iteration.
}
\examples{

adjCombos(letters[1:5])

}
\references{
\url{http://stackoverflow.com/a/20157957/1270695}
}
\author{
Ananda Mahto
}
