% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_flatten.R
\name{col_flatten}
\alias{col_flatten}
\alias{col_flattenLong}
\title{Flatten List Columns into a Wide or Long Form}
\usage{
col_flatten(indt, cols, drop = FALSE)

col_flattenLong(indt, cols)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{cols}{Character vector containing the names of list columns}

\item{drop}{Logical. Should the list columns be dropped from the original 
\code{data.table}?}
}
\value{
A \code{data.table}.
}
\description{
Converts list columns into separate columns or into a long form.
}
\examples{

df <- structure(
  list(CAT = structure(1:2, .Label = c("A", "B"), class = "factor"),
       COUNT = list(1:3, 4:5), TREAT = list(c("Treat-a", "Treat-b"), 
       c("Treat-c", "Treat-d", "Treat-e"))), 
       .Names = c("CAT", "COUNT", "TREAT"), 
       row.names = c(NA, -2L), class = "data.frame")
       
col_flatten(df, c("COUNT", "TREAT"), TRUE)


col_flattenLong(df, c("COUNT", "TREAT"))

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/34206003/1270695}
}

