% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyCalendar.R
\name{dailyCalendar}
\alias{dailyCalendar}
\alias{WeekDays}
\alias{dailyCalendar}
\title{Creates a Calendar in R}
\usage{
WeekDays(startOn = "Monday", abbreviate = FALSE)

dailyCalendar(startDate = Sys.Date(), days = 30, startOn = "Monday",
  fancy = FALSE)
}
\arguments{
\item{startOn}{The day of the week to start on. Defaults to \code{"Monday"}.}

\item{abbreviate}{Logical. Should the result be the abbreviated weekday name?
Defaults to \code{FALSE}.}

\item{startDate}{What should be the first date in the calendar? Defaults to
\code{Sys.Date()}.}

\item{days}{How many days do you want in your calendar? Defaults to \code{30}.}

\item{fancy}{Logical. Should a more nicely formatted version of the calendar
be displayed? Defaults to \code{FALSE}.}
}
\value{
A vector, a \code{data.frame}, or a \code{list}, depending on which function is
called with what arguments.
}
\description{
Creates a daily calendar in R.
}
\examples{
WeekDays()
WeekDays("Thursday", TRUE)
dailyCalendar(startDate = "2013-12-27", days = 10)
dailyCalendar(startDate = "2013-12-27", days = 10, startOn = "Friday")
dailyCalendar(days = 40, fancy = TRUE)
}
\author{
Ananda Mahto
}
