% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupe_thresh.R
\name{dupe_thresh}
\alias{dupe_thresh}
\title{Filters a Vector According to Number of Duplicates}
\usage{
dupe_thresh(invec, count)
}
\arguments{
\item{invec}{The input vector.}

\item{count}{The threshold for duplicates. See "Details".}
}
\value{
A vector.
}
\description{
Filters a vector according to the number of duplicates in the vector, where 
the conditions for the acceptable number of duplicate values are specified.
}
\details{
The \code{"count"} parameter can be either a single digit or a 
character vector showing the desired comparison to be used as the threshold
(for example "> 5"). If no binary relational operator is specified, the 
relational operator used is \code{>=}.
}
\examples{

set.seed(1)
x <- sample(letters[1:10], 35, TRUE)
sort(table(x))

table(dupe_thresh(x, 3))
table(dupe_thresh(x, "<3"))
table(dupe_thresh(x, "== 3"))
table(dupe_thresh(x, "!=3"))

}
\references{
\url{http://stackoverflow.com/q/29973061/1270695}
}
\author{
Ananda Mahto
}
