% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expandRows}
\alias{expandRows}
\title{"Expand" the rows of a \code{data.frame}}
\usage{
expandRows(dataset, count, count.is.col = TRUE, drop = TRUE)
}
\arguments{
\item{dataset}{The input \code{data.frame}.}

\item{count}{The numeric vector of counts OR the column from the
\code{data.frame} that contains the count data.  If \code{count} is a single
digit, it is assumed that all rows should be repeated by this amount.}

\item{count.is.col}{Logical. Is the \code{count} value a column from the
input \code{data.frame}? Defaults to \code{FALSE}.}

\item{drop}{Logical. If \code{count.is.col = TRUE}, should the "count"
column be dropped from the result?  Defaulst to \code{TRUE}.}
}
\value{
A \code{data.frame}.
}
\description{
Expands (replicates) the rows of a \code{data.frame}, either by a fixed
number, a specified vector, or a value contained in one of the columns in
the source \code{data.frame}.
}
\examples{
mydf <- data.frame(x = c("a", "b", "q"), y = c("c", "d", "r"), count = c(2, 5, 3))
mydf
expandRows(mydf, "count")
expandRows(mydf, "count", drop = FALSE)
expandRows(mydf, count = 3)
expandRows(mydf, count = 3, count.is.col = FALSE)
expandRows(mydf, count = c(1, 5, 9), count.is.col = FALSE)
}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/a/19519828/1270695}
}

