% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftable2dt.R
\name{ftable2dt}
\alias{ftable2dt}
\title{Convert an \code{ftable} or an \code{array} Object to a \code{data.table}}
\usage{
ftable2dt(inarray, direction = "wide")
}
\arguments{
\item{direction}{Should the reslut be "wide" (with multiple measurement 
columns) or "long" (with a single measurement column)? Defaults to \code{"wide"}.}

\item{mydata}{The input \code{ftable} or \code{array}}
}
\value{
A \code{data.table}
}
\description{
While convenient methods exist for converting \code{table}s and 
other objects to \code{data.tables}s, such methods do not exist for converting 
an \code{ftable} to a \code{data.table}. An \code{ftable} is essentially a
\code{matrix} with \code{attributes} for the rows and columns, which can be
nested.
}
\examples{

x <- ftable(Titanic, row.vars = 1:3)
x
ftable2dt(x)
ftable2dt(x, direction = "long")

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/a/11143126/1270695}
}

