% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwf2csv.R
\name{fwf2csv}
\alias{fwf2csv}
\title{Creates a CSV Representation of Data Accoding to Stacks of Whitespace}
\usage{
fwf2csv(infile, toDF = FALSE, ...)
}
\arguments{
\item{infile}{The input file. Can also be \code{"clipboard"} to read directly
from the clipboard.}

\item{toDF}{Logical. Should the file be read in while we are at it? Defaults
to \code{FALSE}.}

\item{\dots}{Other arguments to be passed to \code{read.table}.}
}
\value{
A vector or a \code{data.frame}, depending on the value in \code{toDF}.
}
\description{
Uses awk to convert a fixed-width file to a CSV based on stacks of whitespace.
}
\note{
Only tested on Linux.
}
\examples{

myfile <- tempfile(fileext = ".txt")
Lines <- c("aaa  b b ccc      345", "ddd  fgt f u      3456", "e r  der der      5 674")
cat(Lines, sep = "\\n")
cat(Lines, sep = "\\n", file = myfile)
fwf2csv(myfile)
fwf2csv(myfile, TRUE, header = FALSE)
 
}
\references{
\url{http://stackoverflow.com/q/30868600/1270695}
}
\author{
Ananda Mahto and \href{http://stackoverflow.com/users/1745001/ed-morton}{Ed Morton}.
}
