% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_stem.R
\name{grouped_stem}
\alias{grouped_stem}
\alias{print.grouped_stem}
\title{Create a Grouped Stem-and-Leaf Plot}
\usage{
grouped_stem(invec, n = 2)

\method{print}{grouped_stem}(x, ...)
}
\arguments{
\item{invec}{The input vector. This function only works with integers.}

\item{n}{The number of stem values to be grouped}

\item{x}{The object to be printed.}

\item{\dots}{Not used.}
}
\value{
A \code{list} printed with stem-and-leaf formatting
}
\description{
Create a stem-and-leaf plot where the stems can be grouped by multiple values
and the leaves indicate where the values are split
}
\examples{

set.seed(1)
data_pos <- sample(0:50, 100, TRUE)
grouped_stem(data_pos, 2)

data_neg <- sample(-50:-1, 100, TRUE)
grouped_stem(data_neg, 2)

data_pos_neg <- c(0, sample(-50:50, 100, TRUE))
grouped_stem(data_pos_neg, 3)

}
\references{
\url{https://stackoverflow.com/q/62044245/1270695}
}
\seealso{
\code{\link[graphics:stem]{graphics::stem()}}
}
\author{
Ananda Mahto
}
