% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupedMedian.R
\name{GroupedMedian}
\alias{GroupedMedian}
\title{Calculate the Median of Already Grouped Data}
\usage{
GroupedMedian(frequencies, intervals, sep = NULL, trim = NULL)
}
\arguments{
\item{frequencies}{A vector of frequencies.}

\item{intervals}{A 2-row \code{matrix} with the same number of columns as
the length of frequencies, with the first row being the lower class
boundary, and the second row being the upper class boundary. Alternatively,
\code{intervals} may be a column in your \code{data.frame}, and you may
specify \code{sep} (and possibly, \code{trim}) to have the
\code{GroupedMedian} function automatically create the required
\code{matrix} for you.}

\item{sep}{Optional. If the \code{intervals} are represented by a character
vector with a character separating the interval ranges.}

\item{trim}{Characters to trim from the vector before splitting. For
example, if you are doing this on the output of \code{cut} (where, for some
reason, you no longer have access to the original data), you can use the
pre-set trim pattern \code{"cut"}.}
}
\value{
A single numeric value representing the grouped median.
}
\description{
Calculates the median of already grouped data given the interval ranges and
the frequencies of each group.
}
\examples{

mydf <- structure(list(salary = c("1500-1600", "1600-1700", "1700-1800",
        "1800-1900", "1900-2000", "2000-2100", "2100-2200", "2200-2300",
        "2300-2400", "2400-2500"), number = c(110L, 180L, 320L, 460L,
        850L, 250L, 130L, 70L, 20L, 10L)), .Names = c("salary", "number"),
        class = "data.frame", row.names = c(NA, -10L))
mydf

GroupedMedian(frequencies = mydf$number, intervals = mydf$salary, sep = "-")

## Example with intervals manually specified
X <- rbind(c(1500, 1600, 1700, 1800, 1900, 2000, 2100, 2200, 2300, 2400),
           c(1600, 1700, 1800, 1900, 2000, 2100, 2200, 2300, 2400, 2500))

GroupedMedian(mydf$number, X)

set.seed(1)
x <- sample(100, 100, replace = TRUE)
y <- data.frame(table(cut(x, 10)))

GroupedMedian(y$Freq, y$Var1, sep = ",", trim = "cut")

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/a/18931054/1270695}
}

