% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthener.R
\name{lengthener}
\alias{lengthener}
\title{Lengthens a Dataset by Combination of Columns}
\usage{
lengthener(indt, n = 2)
}
\arguments{
\item{indt}{The input dataset.}

\item{n}{The number of columns expected. Passed on to \code{combn}.}
}
\value{
A \code{data.table}.
}
\description{
Creates a long dataset of "n" columns comprising the combinations of values
from different columns.
}
\examples{

mydf <- as.data.frame(matrix(c(1,2,3,4,0,0,1,1), byrow = TRUE, nrow = 2))
lengthener(mydf, 2)
lengthener(mydf, 3)

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/35690478/1270695}
}

