% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_unlister.R
\name{list_unlister}
\alias{list_unlister}
\title{Unlists Columns of Lists by Combinations of Values}
\usage{
list_unlister(indt, addRN = TRUE)
}
\arguments{
\item{indt}{The input \code{data.table}. The input can also be a \code{list}, 
possibly with elements of different lengths.}

\item{addRN}{Logical. Should a column named "rn" be added to the 
\code{data.table}? Such a column is required, so if it does not already exist,
it is suggested to keep this set to \code{TRUE}, the default.}
}
\value{
A \code{data.table}.
}
\description{
Unlists columns of lists by row creating combinations of values in the process.
}
\examples{

L1 <- list(list("A", c("B", "C")), list(1:2, 1:3))
list_unlister(L1)

## Note the NULLs and the shorter length of the first list item
L2 <- list(V1 = list("A", c("A", "B"), "X", NULL),
V2 = list(1, c(1, 2, 3), c(1, 2), c(1, 2, 3, 4), 1),
V3 = list(c("a", "b"), "c", "d", c("e", "f"), c("g", "h", "i")))
list_unlister(L2)

DT <- data.table(x1 = list("A", c("A", "B"), "X", NULL, c("Z", "W")),
x2 = list(1, c(1, 2, 3), c(1, 2), c(1, 2, 3, 4), 1),
x3 = list(c("a", "b"), "c", "d", c("e", "f"), c("g", "h", "i")))
list_unlister(DT)

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/23217958/1270695}
}

