% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_tribble.R
\name{mc_tribble}
\alias{mc_tribble}
\title{Like \code{dput} for Those Confined to the Tidyverse}
\usage{
mc_tribble(indf, indents = 4, mdformat = TRUE)
}
\arguments{
\item{indf}{The input \code{data.frame}.}

\item{indents}{The number of spaces to indent each line of output. Defaults
to \code{4}.}

\item{mdformat}{Logical. Whether or not to add 4 spaces before every line in
order to format as a code block. Defaults to \code{TRUE}.}
}
\description{
Creates a \code{dput}-like pasteable format that can be used to create small tables.
}
\examples{

\dontrun{
short_iris <- head(iris)
mc_tribble(short_iris)
}

}
\references{
\url{http://stackoverflow.com/q/42839626/1270695}
}
\author{
Ananda Mahto. Name courtesy of \href{https://stackoverflow.com/users/1191259/frank}{Frank}.
}
