% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_wide.R
\name{melt_wide}
\alias{melt_wide}
\title{Melt Wide data.tables into Long data.tables}
\usage{
melt_wide(data, id.vars, new.names)
}
\arguments{
\item{data}{The input \code{data.frame}}

\item{id.vars}{ID variables}

\item{new.names}{The new names for the resulting columns}
}
\value{
A long \code{data.table}
}
\description{
Reshapes double and tripple wide \code{data.table}s to long
\code{data.table}s
}
\examples{

triplewide <- structure(list(ID = 1:4,
                             w1d1t1 = c(4L, 3L, 2L, 2L),
                             w1d1t2 = c(5L, 4L, 3L, 3L),
                             w1d2t1 = c(6L, 5L, 5L, 4L),
                             w1d2t2 = c(5L, 4L, 5L, 2L),
                             w2d1t1 = c(6L, 5L, 4L, 3L),
                             w2d1t2 = c(5L, 4L, 5L, 5L),
                             w2d2t1 = c(6L, 3L, 6L, 3L),
                             w2d2t2 = c(7L, 4L, 3L, 2L)),
                             .Names = c("ID", "w1d1t1", "w1d1t2",
                             "w1d2t1", "w1d2t2", "w2d1t1", "w2d1t2",
                             "w2d2t1", "w2d2t2"),
                             class = "data.frame",
                             row.names = c(NA, -4L))
triplewide
triplewide.long <- melt_wide(triplewide, id.vars="ID",
                            new.names=c("week", "day", "trial"))
triplewide.long
data.table::dcast(triplewide.long, ID + week + day ~ trial)

}
\references{
\url{http://stackoverflow.com/a/10170630/1270695}
}
\author{
Ananda Mahto
}
