% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragged.R
\name{ragged}
\alias{ragged}
\alias{print.ragged}
\title{Display a \code{data.frame} with "ragged" keys}
\usage{
ragged(indt, keys, blank = "")

\method{print}{ragged}(x)
}
\arguments{
\item{indt}{The input \code{data.frame} or \code{data.table}}

\item{keys}{The variables to be used as keys or grouping variables}

\item{blank}{The character to print to show nesting. Defaults to "".}

\item{x}{The object to be printed.}
}
\value{
A \code{list} with a "ragged" object and the sorted \code{data.table}. The custom
\code{print} method displays the "ragged" result, but allows further use of \code{data.table}.
}
\description{
This is a display method for \code{data.frame}s to show ragged key/grouping variables,
similar to \code{ftable}
}
\examples{

before= data.frame(C1= c(rep("A", 5), rep("L", 2)),
C2= c("B", rep("E", 3), rep("K", 2), "L"),
C3= c("C", "F", rep("H", 5)),
C4= c("D", "G", "I", rep("J", 4)), 
stringsAsFactors = FALSE)

ragged(before, c("C1", "C2"))
ragged(before, names(before), ":")
data(diamonds, package = "ggplot2")
ragged(head(diamonds, 30), c("cut", "color"), ":")[, mean(price), .(cut, color)]

}
\references{
\url{https://stackoverflow.com/q/41324110/1270695}
}
\seealso{
\code{\link[stats:ftable]{stats::ftable()}}
}
\author{
Ananda Mahto
}
