% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readClip}
\alias{readClip}
\title{Read Clipboard Regardless of OS}
\usage{
readClip()
}
\value{
character string containing text on the clipboard.
}
\description{
Different operating systems have different ways of handling the clipboard.
Given the frequency with which text is copied to the clipboard to place in
an answer on StackOverflow, this utility is provided.
}
