% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_portion.R
\name{replace_portion}
\alias{replace_portion}
\title{Find and Replace a Portion of a Vector}
\usage{
replace_portion(invec, find, replace)
}
\arguments{
\item{invec}{The input vector.}

\item{find}{What is the sequence that you're looking for?}

\item{replace}{What do you want to replace the values with?}
}
\value{
A vector the same length as the input vector.
}
\description{
Searches for a pattern in a vector and replaces it with a provided replacement pattern.
}
\examples{

x <- c(1, 2, 3, 1, 0, 1, 0, 1, 2, 3, 4, 1, 0, 1)
replace_portion(x, c(1, 0, 1), c(9, 9, 9))

}
\author{
Ananda Mahto
}
