% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReshapeLong.R
\name{ReshapeLong}
\alias{ReshapeLong}
\title{Reshape Data into a Semi-Long Format}
\usage{
ReshapeLong(indt, stubs, sep = NULL)
}
\arguments{
\item{indt}{The input \code{data.table}.}

\item{stubs}{Character vector containing the uniquely identifying stub 
portion of the variable names.}

\item{sep}{Not presently used.}
}
\value{
A \code{data.table}.
}
\description{
Reshapes data with multiple measurements in a wide format into 
a long format with one column per measurement type.
}
\examples{

\dontrun{
library(foreign)
dadmom <- read.dta("http://www.ats.ucla.edu/stat/stata/modules/dadmomw.dta")
ReshapeLong(dadmom, c("name", "inc"))
}

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/10468969/1270695}
}

