% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shifter}
\alias{shifter}
\title{"Shift" the values of a vector ahead or behind by a specified amount}
\usage{
shifter(x = 1:10, n = 1)
}
\arguments{
\item{x}{The range that you are shifting}

\item{n}{How much of a shift you want}
}
\value{
A vector of shifted values
}
\description{
This function "shifts" the values of a vector by a specified amount. For
instance, if you are starting with a vector, "x", where the range of values
is between 1 and 10, and you want 10 to be replaced by 9, 9 to be replaced
by 8, and so on, with 1 being ultimately replaced by 10, this funciton
should be of use.
}
\examples{
set.seed(1)
X <- sample(10, 20, replace = TRUE)
X

shifter()[X]
shifter(n = -2)[X]
}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/a/20825012/1270695}
}

