% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabulateInt.R
\name{TabulateInt}
\alias{TabulateInt}
\title{Allows Tabulate to Be Used with Negative Integers Too}
\usage{
TabulateInt(vec)
}
\arguments{
\item{vec}{The input vector}
}
\value{
A named integer vector. There is a bin for each of the values 1, ..., nbins.
}
\description{
Modifies tabulate to work with non-positive integers too.
}
\note{
The behavior on non-integers might be somewhat unpredictable, but should 
be somewhat like using `table(cut(...))` with breaks being from the minimum
to the maximum + 1. See the "Examples" section.
}
\examples{

x <- c(-5, -5, 3, 1, 0, 2, 5, -4, 0, 0, 1)
TabulateInt(x)

## Compare
tabulate(x)
table(x)
table(factor(x, min(x):max(x)))

## Non-integers
set.seed(1)
x <- rnorm(20)
TabulateInt(x)
table(cut(x, seq(min(x), max(x)+1, 1), include.lowest = TRUE))

}
\author{
Ananda Mahto
}

