% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_by_row.R
\name{unlist_by_row}
\alias{unlist_by_row}
\alias{unlist_by_row}
\alias{unlist_by_col}
\title{Unlists the Values in a Rectangular Dataset by Row or Column}
\usage{
unlist_by_row(indt, source = TRUE)

unlist_by_col(indt, source = TRUE)
}
\arguments{
\item{indt}{The input dataset.}

\item{source}{Logical. Should columns indicating the original row and column
positions be returned. Defaults to \code{TRUE}.}
}
\value{
A \code{data.table} if \code{source = TRUE} or a vector.
}
\description{
Unlists the values in a rectangular dataset (like a \code{matrix},
\code{data.frame}, or \code{data.table}) by row.
}
\examples{

unlist_by_row(mtcars)


unlist_by_col(mtcars)

}
\references{
\url{http://stackoverflow.com/q/36073947/1270695}
}
\author{
Ananda Mahto
}
