% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorBind.R
\name{vectorBind}
\alias{vectorBind}
\title{Bind Vectors Column-Wise According to Name}
\usage{
vectorBind(...)
}
\arguments{
\item{\dots}{The objects that need to be combined.}
}
\value{
A matrix.
}
\description{
Combines named vectors into a matrix with the rows being the names of the
vector elements, and the columns being the name of the source vector.
}
\examples{

set.seed(1)
t1 <- table(sample(LETTERS[c(1, 2, 4)], 20, TRUE))
t2 <- table(sample(LETTERS[c(1, 2, 3)], 20, TRUE))
t3 <- table(sample(LETTERS[c(2, 4, 5)], 20, TRUE))

vectorBind(t1, t2, t3)

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/25639223/1270695}
}

