% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_value.R
\name{word_value}
\alias{word_value}
\title{Converts a Vector of Words to Numeric Values Based on Letter Positions}
\usage{
word_value(words, dtOut = FALSE)
}
\arguments{
\item{words}{The input vector of words.}

\item{dtOut}{Logical. Should the output be a \code{data.table} comprising
the words and the values. Defaults to \code{FALSE}.}
}
\value{
A named numeric vector or a \code{data.table}.
}
\description{
Uses the numeric position of the letters in a word to create a numeric value
for a word.
}
\details{
The function converts the input to lowercase, removes anything that
is not between the letters "a" and "z", and transliterates accented characters
to their ASCII equivalent before converting the word to a numeric value.
}
\examples{

myvec <- c("and", "dad", "cat", "fox", "mom", 
           "add", "dan", "naive", "non-descript")
word_value(myvec)
word_value(myvec, TRUE)

}
\references{
\url{http://stackoverflow.com/q/36097446/1270695}
}
\author{
Ananda Mahto and \href{http://stackoverflow.com/users/640595/jota}{Jota}.
}
