% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeClip}
\alias{writeClip}
\title{Write to Clipboard on Multiple OSes}
\usage{
writeClip(object)
}
\arguments{
\item{object}{character. Character to be copied to the clipboard}
}
\value{
Returns nothing to R. Returns character string to the clipboard
}
\description{
This function works on Windows, Mac and Linux. It copies a 
character string or vector of characters to the clipboard and interprets
a vector of characters as one character with each element being newline
separated. If using Linux, xclip is used as the clipboard. So for the
function to work, xclip must be installed.
}
\details{
If using Linux, xclip will be used as the clipboard. To paste from 
xclip, either use middle click or the command \code{xclip -o} in the shell.
}
