% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this_by_n.R
\name{this_by_n}
\alias{this_by_n}
\title{Apply This By Every n Values}
\usage{
this_by_n(invec, n = 3, FUN = sum, pad_val = NA)
}
\arguments{
\item{invec}{The input vector.}

\item{n}{By how many values?}

\item{FUN}{The function to apply to each set of n values.}

\item{pad_val}{The value to padd the resulting vector with. Defaults to \code{NA}.}
}
\value{
A vector the same length as the input vector.
}
\description{
Applies a function by every n values to a vector minus the first value.
}
\examples{

x <- c(1, 2, 3, 4, 7, 9, 2, 4)
this_by_n(x, 3, mean)
this_by_n(x, 2, max)
this_by_n(x, 4, min)

}
\author{
Ananda Mahto
}
\references{
\url{http://stackoverflow.com/q/34563693/1270695}
}

